initial cleanup/refactor
