+++ src/lib/components/form-checkbox.svelte
@@ -1,17 +1,9 @@
 <script lang="ts">
-	import { createCheckboxAction, getCtx } from "@/lib/internal/index.js";
+	import { getFormField } from "@/lib/index.js";
 	import type { CheckboxProps } from "../types.js";
 
 	type $$Props = CheckboxProps;
-
-	const { ids, value, name, attrStore } = getCtx();
-
-	const action = createCheckboxAction({
-		id: ids.input,
-		value,
-		name,
-		attrs: attrStore
-	});
+	const { actions } = getFormField();
 </script>
 
-<input type="checkbox" use:action />
+<input type="checkbox" use:actions.checkbox />
