+++ src/lib/components/form-description.svelte
@@ -1,19 +1,12 @@
 <script lang="ts">
-	import { createDescriptionAction, getCtx } from "@/lib/internal/index.js";
+	import { getFormField } from "@/lib/index.js";
 	import type { DescriptionProps } from "../types.js";
 
 	type $$Props = DescriptionProps;
-
 	export let tag = "p";
-
-	const { ids, hasDescription } = getCtx();
-
-	const action = createDescriptionAction({
-		id: ids.description,
-		hasDescription
-	});
+	const { actions } = getFormField();
 </script>
 
-<svelte:element this={tag} use:action {...$$restProps}>
+<svelte:element this={tag} use:actions.description {...$$restProps}>
 	<slot />
 </svelte:element>
