+++ src/lib/components/form-radio.svelte
@@ -1,17 +1,9 @@
 <script lang="ts">
-	import { createRadioAction, getCtx } from "@/lib/internal/index.js";
+	import { getFormField } from "@/lib/index.js";
 	import type { RadioProps } from "../types.js";
 
 	type $$Props = RadioProps;
-
-	const { ids, value, name, attrStore } = getCtx();
-
-	const action = createRadioAction({
-		id: ids.input,
-		value,
-		name,
-		attrs: attrStore
-	});
+	const { actions } = getFormField();
 </script>
 
-<input type="radio" {...$$restProps} use:action />
+<input type="radio" {...$$restProps} use:actions.radio />
