+++ src/lib/helpers/get-form-field.ts
@@ -1,0 +1,8 @@
+import { getContext } from "svelte";
+import { FORM_FIELD_CONTEXT, type FieldContext } from "@/lib/internal/index.js";
+
+export function getFormField() {
+	return getContext<FieldContext>(FORM_FIELD_CONTEXT);
+}
+
+export type { FieldContext };
