+++ src/lib/index.ts
@@ -1,5 +1,5 @@
 // Reexport your entry components here
-import { getCtx } from "./internal/index.js";
+import { getFormField } from "./internal/index.js";
 export * as Form from "./components/index.js";
 export * from "./types.js";
-export { getCtx as getFormField };
+export { getFormField };
