+++ src/lib/internal/index.ts
@@ -1,8 +1,13 @@
-export * from "./form-field.js";
+export * from "./form-field/create-field.js";
 export * from "./types.js";
-export * from "./actions.js";
+export * from "./form-field/create-actions.js";
 export * from "./callbacks.js";
-export * from "./actions.js";
+export * from "./form-field/create-actions.js";
 export * from "./events.js";
 export * from "./is.js";
 export * from "./derived.js";
+export * from "./form-field/create-handlers.js";
+export * from "./form-field/create-ids.js";
+export * from "./form/index.js";
+export * from "./dom.js";
+export * from "./form-field/index.js";
